Add @property decorator to Response.next()
In https://github.com/python/typeshed/commit/bd7222c39669cd166f9885318b3ede779e7c2edc , an annotated stub was added for the `Response.next()` method. In the requests library, this method has the `@property` decorator applied to it; I believe that the stub should also have that decorator applied.

This is my first time making a PR to this repository, please let me know if I've made a mistake!